+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "js-syntax-extension" extension will be documented in
 - Added invert if quick action refactor.
   - Only invert if the last curly brace is properly formatted. The tabs of if must match tabs for }
 - Fix bug with enums not generating correctly.
+- Add extract function code action.
 
 # 2.6.0+